requirements.txt
asgiref==3.11.0
celery==5.4.0
dj-database-url==3.1.0
Django==5.2.10
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
Pillow==12.1.0
PyJWT==2.10.1
redis==5.2.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
